exploded tests into a test/ dir
